Extra csv data files for consideration
I noticed that the only available data seems to be a ship register for the titanic? So I took to kaggle and selected 3 datasets for us to choose from. I performed a very basic exploration of each dataset and I have suggested a dataset for us to use based on my findings in the conclusion. CSVs were too large to upload, but links are in the txt doc. Please read data_exploration.txt for more details.